pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "backlog-automation"
version = "0.1.0"
description = "Automated backlog creation from Excel/Word into Jira, Confluence, and Figma"
requires-python = ">=3.11"
dependencies = [
    # Input parsing
    "pandas>=2.2",
    "openpyxl>=3.1",
    "python-docx>=1.1",
    # Atlassian integrations
    "atlassian-python-api>=3.41",
    # Figma REST API
    "httpx>=0.27",
    # AI-powered story generation
    "anthropic>=0.30",
    # Workflow orchestration
    "prefect>=3.0",
    # Config and secrets
    "pydantic-settings>=2.3",
    "python-dotenv>=1.0",
]

[project.scripts]
backlog = "backlog_automation.main:app"

[tool.hatch.build.targets.wheel]
packages = ["src/backlog_automation"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.5",
]
